medicine_ordering_system/requirement.txt
amqp==5.3.1
asgiref==3.9.1
billiard==4.2.1
celery==5.5.3
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
contourpy==1.3.3
crispy-bootstrap5==2025.6
cycler==0.12.1
Cython==3.1.3
Django==5.2.6
django-anymail==13.1
django-cors-headers==4.7.0
django-crispy-forms==2.4
django-extensions==4.1
django-redis==6.0.0
django-storages==1.14.6
djangorestframework==3.16.1
drf-yasg==1.21.10
fonttools==4.59.2
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
joblib==1.5.2
kiwisolver==1.4.9
kombu==5.5.4
matplotlib==3.10.6
mysqlclient==2.2.7
narwhals==2.3.0
numpy==2.3.2
packaging==25.0
pandas==2.3.2
patsy==1.0.1
pillow==11.3.0
plotly==6.3.0
pmdarima==2.0.4
prompt_toolkit==3.0.52
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
redis==6.4.0
requests==2.32.5
scikit-learn==1.7.1
scipy==1.16.1
seaborn==0.13.2
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
statsmodels==0.14.5
threadpoolctl==3.6.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
